posformer_requirements.txt
pytorch-lightning>=1.5.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.20.0
tqdm>=4.50.0
matplotlib>=3.3.0
typer>=0.4.0
einops>=0.6.0